requirements.txt
appJar==0.94.0
backcall==0.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.2.2
coloredlogs==10.0
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
future==0.17.1
getch==1.0
glfw==1.8.3
gym==0.14.0
humanfriendly==4.18
idna==2.8
imageio==2.5.0
install==1.3.4
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
joblib==1.0.1
kiwisolver==1.1.0
llvmlite==0.34.0
lockfile==0.12.2
lxml==4.4.1
malmo==0.36.0.0
malmoenv==0.0.8
-e git+https://github.com/Microsoft/malmo-challenge@a1dec75d0eb4cccc91f9d818a4ecae5fa1ac906f#egg=malmopy
matplotlib==3.0.3
minerl==0.2.9
numba==0.51.2
numpy==1.17.2
opencv-python==4.1.1.26
pandas==0.25.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
prompt-toolkit==2.0.10
psutil==5.6.3
ptyprocess==0.6.0
pycparser==2.19
pygame==1.9.6
pyglet==1.3.2
Pygments==2.4.2
pyparsing==2.4.2
Pyro4==4.77
pyspellchecker==0.5.5
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
scikit-learn==0.24.1
scipy==1.3.1
serpent==1.28
Shapely==1.7.1
six==1.12.0
sklearn==0.0
soupsieve==2.0.1
thesaurus==0.1
thesaurus-cli==1.0.1
threadpoolctl==2.1.0
tqdm==4.36.1
traitlets==4.3.3
typing==3.7.4.1
urllib3==1.25.6
wcwidth==0.1.7
